requirements.txt
fastapi
uvicorn[standard]
chromadb
sentence-transformers
transformers
torch
torchvision
torchaudio
faster-whisper
python-docx
pdfplumber
numpy
tqdm
Pillow
pytesseract
faiss-cpu
llama-cpp-python
python-multipart
aiofiles
python-dotenv
scipy
soundfile
